+++ litex/soc/software/include/hw/common.h
@@ -14,34 +14,34 @@
 #ifdef __ASSEMBLER__
 #define MMPTR(x) x
 #else /* ! __ASSEMBLER__ */
-#define MMPTR(x) (*((volatile unsigned int *)(x)))
+#define MMPTR(x) (*((volatile unsigned long *)(x)))
 
-static inline void csr_writeb(uint8_t value, uint32_t addr)
+static inline void csr_writeb(uint8_t value, unsigned long addr)
 {
 	*((volatile uint8_t *)addr) = value;
 }
 
-static inline uint8_t csr_readb(uint32_t addr)
+static inline uint8_t csr_readb(unsigned long addr)
 {
 	return *(volatile uint8_t *)addr;
 }
 
-static inline void csr_writew(uint16_t value, uint32_t addr)
+static inline void csr_writew(uint16_t value, unsigned long addr)
 {
 	*((volatile uint16_t *)addr) = value;
 }
 
-static inline uint16_t csr_readw(uint32_t addr)
+static inline uint16_t csr_readw(unsigned long addr)
 {
 	return *(volatile uint16_t *)addr;
 }
 
-static inline void csr_writel(uint32_t value, uint32_t addr)
+static inline void csr_writel(uint32_t value, unsigned long addr)
 {
 	*((volatile uint32_t *)addr) = value;
 }
 
-static inline uint32_t csr_readl(uint32_t addr)
+static inline uint32_t csr_readl(unsigned long addr)
 {
 	return *(volatile uint32_t *)addr;
 }
